output/main/requirements.txt
absl-py==0.13.0
altgraph==0.17
asgiref==3.4.1
attrs==21.2.0
auto-py-to-exe==2.9.0
bottle==0.12.19
bottle-websocket==0.2.9
caer==1.9.8
cffi==1.14.5
cycler==0.10.0
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
importlib-metadata==4.8.1
imutils==0.5.4
install==1.3.4
kiwisolver==1.3.2
matplotlib==3.4.2
mediapipe==0.8.6
mutagen==1.45.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.5
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas==1.3.5
pefile==2021.5.24
Pillow==8.3.1
protobuf==3.17.3
pycparser==2.20
pygame==2.1.2
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySide6==6.1.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pywin32-ctypes==0.2.0
shiboken6==6.1.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
whichcraft==0.6.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
